backend/requirements.txt
apispec==6.10.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.5
dnspython==2.8.0
email-validator==2.3.0
Faker==40.8.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
flask-smorest==0.45.0
Flask-WTF==1.2.2
google-auth==2.49.1
google-auth-oauthlib==1.3.0
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
marshmallow==3.23.2
oauthlib==3.3.1
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyee==13.0.1
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-dotenv==1.2.2
python-slugify==8.0.4
requests==2.32.5
requests-oauthlib==2.0.0
SQLAlchemy==2.0.48
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
webargs==8.7.1
Werkzeug==3.1.6
WTForms==3.2.1

